Move BillingInformationListener to the Billing module
