+++ parking/private/parking_factory.cpp
@@ -12,7 +12,7 @@ namespace Parking {
 std::unique_ptr<Parking> ParkingFactory::create(std::unique_ptr<Payments::PaymentSystem> paymentSystem,
                                                 std::unique_ptr<GateControlStrategy> gateControlStrategy,
                                                 Billing::BillingSystem& billingSystem,
-                                                BillingInformationListener& billingInformationListener) {
+                                                Billing::BillingInformationListener& billingInformationListener) {
     return std::make_unique<ParkingImpl>(std::move(paymentSystem), std::move(gateControlStrategy), billingSystem,
                                          billingInformationListener);
 }
