+++ parking/private/pay_on_ticket_machine_strategy.h
@@ -12,6 +12,7 @@
 #include <vector>
 
 namespace Billing {
+class BillingInformationListener;
 class BillingSystem;
 } // namespace Billing
 
@@ -21,13 +22,11 @@ class GatesController;
 
 namespace Parking {
 
-class BillingInformationListener;
-
 class PayOnTicketMachineStrategy: public GateControlStrategy{
 public:
     PayOnTicketMachineStrategy(Billing::BillingSystem& billingSystem,
                                std::unordered_map<std::string, unsigned int>& carsRegistry,
-                               BillingInformationListener& billingListener, Gates::GatesController& gateController);
+                               Billing::BillingInformationListener& billingListener, Gates::GatesController& gateController);
 
     void onCarEntering(std::size_t gateId, const std::string& carId, unsigned int tickId) override;
     void onCarLeaving(std::size_t gateId, const std::string& carId, unsigned int tickId) override;
@@ -38,7 +37,7 @@ private:
     std::unordered_set<std::string> _payedCars;
 
     Billing::BillingSystem& _billingSystem;
-    BillingInformationListener& _billingListener;
+    Billing::BillingInformationListener& _billingListener;
 
     Gates::GatesController& _gateController;
 };
